requirements.txt
# Core orchestration
langchain>=0.0.300
ollama>=0.1.0
langchain-community

# Vector store backends
chromadb>=0.3.24
faiss-cpu>=1.7.3

# Embeddings
sentence-transformers>=2.2.2
tiktoken>=0.4.0

# Document loading
pypdf>=3.8.0
pdfplumber>=0.10.0
unstructured>=0.7.1

# Utility
python-dotenv>=0.21.0
requests>=2.28.1
